Separated community template form
