+++ app/client/src/pages/Editor/PublishCommunityTemplate/index.tsx
@@ -2,64 +2,16 @@ import {
   COMMUNITY_TEMPLATES,
   createMessage,
 } from "@appsmith/constants/messages";
-import { Button, Text } from "design-system";
-import React, { useMemo, useState } from "react";
+import { Text } from "design-system";
+import React from "react";
 import BackButton from "../DataSourceEditor/BackButton";
+import CommunityTemplateForm from "./CommunityTemplateForm";
 import {
-  PublishPageAppSettingContainer,
-  PublishPageBodyContainer,
-  PublishPageFooterContainer,
   PublishPageHeader,
   PublishPageHeaderContainer,
-  PublishPageTemplateDetailsInputContainer,
 } from "./styledComponents";
-import TemplateCardPreview from "./components/TemplateCardPreview";
-import TemplateInfoForm from "./components/TemplateInfoForm";
-import PublishedInfo from "./components/PublishedInfo";
-import AuthorDetailsInput from "./components/AuthorDetailsInput";
-import { getCurrentUser } from "selectors/usersSelectors";
-import { useDispatch, useSelector } from "react-redux";
-import ApplicationSettings from "./components/ApplicationSettings";
-import { publishCommunityTemplate } from "actions/communityTemplateActions";
-import { isPublishingCommunityTempalteSelector } from "selectors/communityTemplatesSelector";
 
 const PublishCommunityTemplate = () => {
-  const currentUser = useSelector(getCurrentUser);
-  const dispatch = useDispatch();
-  const isPublishing = useSelector(isPublishingCommunityTempalteSelector);
-
-  const [templateName, setTemplateName] = useState("");
-  const [templateExcerpt, setTemplateExcerpt] = useState("");
-  const [templateDescription, setTemplateDescription] = useState("");
-  const [templateUseCases, settemplateUseCases] = useState<string[]>([]);
-
-  const [authorName, setAuthorName] = useState(currentUser?.name || "");
-  const [authorEmail, setAuthorEmail] = useState(currentUser?.email || "");
-
-  const [isPublicSetting, setIsPublicSetting] = useState(true);
-  const [isForkableSetting, setIsForkableSetting] = useState(true);
-
-  const isFormValid = useMemo(() => {
-    const requiredFields = [templateName, authorName, authorEmail];
-    const areRequiredFieldsPresent = requiredFields.every(
-      (field) => field.length > 0,
-    );
-    const areSettingsTurnedON = isPublicSetting && isForkableSetting;
-    return areRequiredFieldsPresent && areSettingsTurnedON;
-  }, [
-    templateName,
-    authorName,
-    authorEmail,
-    isPublicSetting,
-    isForkableSetting,
-  ]);
-
-  const publishToCommunity = () => {
-    if (!isFormValid) {
-      return;
-    }
-    dispatch(publishCommunityTemplate());
-  };
   return (
     <>
       <PublishPageHeaderContainer>
@@ -70,54 +22,7 @@ const PublishCommunityTemplate = () => {
           {createMessage(COMMUNITY_TEMPLATES.publishFormPage.title)}
         </Text>
       </PublishPageHeaderContainer>
-      <PublishPageBodyContainer>
-        <TemplateCardPreview
-          excerpt={templateExcerpt}
-          templateName={templateName}
-          useCases={["Operations", "DevOps"]}
-        />
-        <PublishPageTemplateDetailsInputContainer>
-          <TemplateInfoForm
-            setTemplateDescription={setTemplateDescription}
-            setTemplateExcerpt={setTemplateExcerpt}
-            setTemplateName={setTemplateName}
-            setTemplateUseCases={settemplateUseCases}
-            templateDescription={templateDescription}
-            templateExcerpt={templateExcerpt}
-            templateName={templateName}
-            templateUseCases={templateUseCases}
-          />
-          <AuthorDetailsInput
-            authorEmail={authorEmail}
-            authorName={authorName}
-            disableEmail={!!currentUser?.email}
-            disableName={!!currentUser?.name}
-            setAuthorEmail={setAuthorEmail}
-            setAuthorName={setAuthorName}
-          />
-          <PublishPageAppSettingContainer>
-            <ApplicationSettings
-              isForkable={isForkableSetting}
-              isPublic={isPublicSetting}
-              setIsForkable={setIsForkableSetting}
-              setIsPublic={setIsPublicSetting}
-            />
-          </PublishPageAppSettingContainer>
-          <PublishedInfo />
-        </PublishPageTemplateDetailsInputContainer>
-      </PublishPageBodyContainer>
-      <PublishPageFooterContainer>
-        <Button
-          isDisabled={!isFormValid}
-          isLoading={isPublishing}
-          onClick={publishToCommunity}
-          size="md"
-        >
-          {createMessage(
-            COMMUNITY_TEMPLATES.publishFormPage.footerPublishButton,
-          )}
-        </Button>
-      </PublishPageFooterContainer>
+      <CommunityTemplateForm />
     </>
   );
 };
